Drive function changed
In the past, our drive function fed our motors values as high as 255 and
as low as -254, though the motors only care about magnitude of numbers
between -100 and 100, but now our dillema is fixed.  The new drive
funtion will allow for a maximum value of 100 and minimum value of -100.
It also spits out the numbers exponentially, so if a driver wants to
make the robot very slow, he pushes the joystick only a little, much
like on a mouse.